Reduce amount of commits needed by excluding overwriting changes
Reduced the amount of commits by being smarter about what changes to
apply.

The bot now goes over all commits in reverse order (i.e. starting with
the latest commit) and only pushes a changes when a file has not been
processed before.

Note that the file is only labelled as processed
(`processed.push(file)`) as the last task for each file. For modified
files this is required because otherwise the "add-if-pdf-not-found" (see
commit 2a7de66) feature does not work.